requirements.txt
psycopg2-binary
boto3
python-decouple